requirements.txt
python-dotenv==1.0.1
python-dateutil==2.9.0
mistralai<1.0.0
yt-dlp
instaloader
playwright
python-telegram-bot
apscheduler
pytz
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
